requirements.txt
fastapi
uvicorn
extcolors
python-multipart
Pillow
